@@ -1,27 +1,27 @@
 {
-  "_from": "ansi-regex@^4.1.0",
-  "_id": "ansi-regex@4.1.0",
+  "_from": "ansi-regex@^5.0.1",
+  "_id": "ansi-regex@5.0.1",
   "_inBundle": false,
-  "_integrity": "sha512-1apePfXM1UOSqw0o9IiFAovVz9M5S1Dg+4TrDwfMewQ6p/rmMueb7tWZjQ1rx4Loy1ArBggoqGpfqqdI4rondg==",
+  "_integrity": "sha512-quJQXlTSUGL2LH9SUXo8VwsY4soanhgo6LNSm84E1LBcE8s3O0wpdiRzyR9z/ZZJMlMWv37qOOb9pdJlMUEKFQ==",
   "_location": "/ansi-regex",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "ansi-regex@^4.1.0",
+    "raw": "ansi-regex@^5.0.1",
     "name": "ansi-regex",
     "escapedName": "ansi-regex",
-    "rawSpec": "^4.1.0",
+    "rawSpec": "^5.0.1",
     "saveSpec": null,
-    "fetchSpec": "^4.1.0"
+    "fetchSpec": "^5.0.1"
   },
   "_requiredBy": [
     "/strip-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-4.1.0.tgz",
-  "_shasum": "8b9f8f08cf1acb843756a839ca8c7e3168c51997",
-  "_spec": "ansi-regex@^4.1.0",
-  "_where": "/Users/gojiboy/Desktop/Projects/atmos/server/node_modules/strip-ansi",
+  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-5.0.1.tgz",
+  "_shasum": "082cb2c89c9fe8659a311a53bd6a4dc5301db304",
+  "_spec": "ansi-regex@^5.0.1",
+  "_where": "C:\\Users\\jento\\Desktop\\Projects\\practice projects\\atmos-project\\server\\node_modules\\strip-ansi",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -34,14 +34,16 @@
   "deprecated": false,
   "description": "Regular expression for matching ANSI escape codes",
   "devDependencies": {
-    "ava": "^0.25.0",
-    "xo": "^0.23.0"
+    "ava": "^2.4.0",
+    "tsd": "^0.9.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/chalk/ansi-regex#readme",
   "keywords": [
@@ -78,8 +80,8 @@
     "url": "git+https://github.com/chalk/ansi-regex.git"
   },
   "scripts": {
-    "test": "xo && ava",
+    "test": "xo && ava && tsd",
     "view-supported": "node fixtures/view-codes.js"
   },
-  "version": "4.1.0"
+  "version": "5.0.1"
 }
