@@ -1,27 +1,27 @@
 {
-  "_from": "is-fullwidth-code-point@^2.0.0",
-  "_id": "is-fullwidth-code-point@2.0.0",
+  "_from": "is-fullwidth-code-point@^3.0.0",
+  "_id": "is-fullwidth-code-point@3.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-o7MKXE8ZkYMWeqq5O+764937ZU8=",
+  "_integrity": "sha512-zymm5+u+sCsSWyD9qNaejV3DFvhCKclKdizYaJUuHA83RLjb7nSuGnddCHGv0hk+KY7BMAlsWeK4Ueg6EV6XQg==",
   "_location": "/is-fullwidth-code-point",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "is-fullwidth-code-point@^2.0.0",
+    "raw": "is-fullwidth-code-point@^3.0.0",
     "name": "is-fullwidth-code-point",
     "escapedName": "is-fullwidth-code-point",
-    "rawSpec": "^2.0.0",
+    "rawSpec": "^3.0.0",
     "saveSpec": null,
-    "fetchSpec": "^2.0.0"
+    "fetchSpec": "^3.0.0"
   },
   "_requiredBy": [
-    "/ansi-align/string-width"
+    "/string-width"
   ],
-  "_resolved": "https://registry.npmjs.org/is-fullwidth-code-point/-/is-fullwidth-code-point-2.0.0.tgz",
-  "_shasum": "a3b30a5c4f199183167aaab93beefae3ddfb654f",
-  "_spec": "is-fullwidth-code-point@^2.0.0",
-  "_where": "/Users/gojiboy/Desktop/Projects/atmos/server/node_modules/ansi-align/node_modules/string-width",
+  "_resolved": "https://registry.npmjs.org/is-fullwidth-code-point/-/is-fullwidth-code-point-3.0.0.tgz",
+  "_shasum": "f116f8064fe90b3f7844a38997c0b75051269f1d",
+  "_spec": "is-fullwidth-code-point@^3.0.0",
+  "_where": "C:\\Users\\jento\\Desktop\\Projects\\practice projects\\atmos-project\\server\\node_modules\\string-width",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -34,14 +34,16 @@
   "deprecated": false,
   "description": "Check if the character represented by a given Unicode code point is fullwidth",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.3.1",
+    "tsd-check": "^0.5.0",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/sindresorhus/is-fullwidth-code-point#readme",
   "keywords": [
@@ -51,9 +53,7 @@
     "width",
     "unicode",
     "character",
-    "char",
     "string",
-    "str",
     "codepoint",
     "code",
     "point",
@@ -68,10 +68,7 @@
     "url": "git+https://github.com/sindresorhus/is-fullwidth-code-point.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd-check"
   },
-  "version": "2.0.0",
-  "xo": {
-    "esnext": true
-  }
+  "version": "3.0.0"
 }
