@@ -1,27 +1,28 @@
 {
-  "_from": "strip-ansi@^5.1.0",
-  "_id": "strip-ansi@5.2.0",
+  "_from": "strip-ansi@^6.0.1",
+  "_id": "strip-ansi@6.0.1",
   "_inBundle": false,
-  "_integrity": "sha512-DuRs1gKbBqsMKIZlrffwlug8MHkcnpjs5VPmL1PAh+mA30U0DTotfDZ0d2UUsXpPmPmMMJ6W773MaA3J+lbiWA==",
+  "_integrity": "sha512-Y38VPSHcqkFrCpFnQ9vuSXmquuv5oXOKpGeT6aGrr3o3Gc9AlVa6JBfUSOCnbxGGZF+/0ooI7KrPuUSztUdU5A==",
   "_location": "/strip-ansi",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "strip-ansi@^5.1.0",
+    "raw": "strip-ansi@^6.0.1",
     "name": "strip-ansi",
     "escapedName": "strip-ansi",
-    "rawSpec": "^5.1.0",
+    "rawSpec": "^6.0.1",
     "saveSpec": null,
-    "fetchSpec": "^5.1.0"
+    "fetchSpec": "^6.0.1"
   },
   "_requiredBy": [
-    "/ansi-align/string-width"
+    "/string-width",
+    "/wrap-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-5.2.0.tgz",
-  "_shasum": "8c9a536feb6afc962bdfa5b104a5091c1ad9c0ae",
-  "_spec": "strip-ansi@^5.1.0",
-  "_where": "/Users/gojiboy/Desktop/Projects/atmos/server/node_modules/ansi-align/node_modules/string-width",
+  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-6.0.1.tgz",
+  "_shasum": "9e26c63d30f53443e9489495b2105d37b67a85d9",
+  "_spec": "strip-ansi@^6.0.1",
+  "_where": "C:\\Users\\jento\\Desktop\\Projects\\practice projects\\atmos-project\\server\\node_modules\\string-width",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -32,17 +33,17 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-regex": "^4.1.0"
+    "ansi-regex": "^5.0.1"
   },
   "deprecated": false,
   "description": "Strip ANSI escape codes from a string",
   "devDependencies": {
-    "ava": "^1.3.1",
-    "tsd-check": "^0.5.0",
-    "xo": "^0.24.0"
+    "ava": "^2.4.0",
+    "tsd": "^0.10.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=8"
   },
   "files": [
     "index.js",
@@ -80,7 +81,7 @@
     "url": "git+https://github.com/chalk/strip-ansi.git"
   },
   "scripts": {
-    "test": "xo && ava && tsd-check"
+    "test": "xo && ava && tsd"
   },
-  "version": "5.2.0"
+  "version": "6.0.1"
 }
