@@ -1,44 +1,28 @@
-import { useDispatch, useSelector } from "react-redux"
-import { contentActions } from "../store/Content"
-import { useHistory } from "react-router";
-import ConfModal from './CModal';
-import Backdrop from './Backdrop'
+import { Link } from "react-router-dom";
 import MusicForm from './MusicForm';
 import VideoForm from './VideoForm';
-import VideoContent from "./VideoContent";
+import { useSelector } from "react-redux";
 import { AnimationWrapper } from 'react-hover-animation'
-import styles from './welcomeStyles.module.css'
+import styles from './MainMenuStyles.module.css'
 
 
 function MainMenu(props) {
-    const history = useHistory()
-    const dispatch = useDispatch()
-    const showModal = useSelector((state) => state.content.showCreateModal)
-
-
-    function useContinueContent() {
-        history.push("/dashboard")
-    }
-
-    function toggleModalHandler() {
-        dispatch(contentActions.showCreateModal())
-    }
+    const videoChoice = useSelector((state) => state.youtube.currentChoice)
+    const musicChoice = useSelector((state) => state.content.musicButtonChoice)
 
     return (
-        <div className="pracmenu">
-            <h2>{props.text}</h2>
-            <div className="formContainer my-3">
+        <div className={styles.mainMenu}>
+            <h2>Please select your content</h2>
+            <div className={styles.formContainer}>
                 <MusicForm />
                 <VideoForm />
             </div>
-            <div className="pracactions">
+            {(videoChoice !== "none" && musicChoice !== "none") && <div>
                 <AnimationWrapper>
-               
-             <button className="my-5" className={styles.buttonStyles} onClick={toggleModalHandler}>Create Atmosphere</button>
-            </AnimationWrapper></div>
-            {showModal && <ConfModal onCancel={toggleModalHandler} onConfirm={useContinueContent} />}
-            {showModal && <Backdrop onCancel={toggleModalHandler} />}
-            {/* <VideoContent /> */}
+                    <Link to="/dashboard" className="m-3 btn btn-success">Create Atmos
+                    </Link>
+                </AnimationWrapper>
+            </div>}
         </div>
     )
 }
