@@ -1,7 +1,6 @@
 // import { useEffect } from "react"
-import { useDispatch } from "react-redux"
-import { useSelector } from "react-redux"
-import { controlActions } from "../store/UserControls"
+// import { useSelector } from "react-redux"
+
 import SpotifyMusicPlayer from "./SpotifyPlayer"
 // import SpotifyWebApi from "spotify-web-api-node"
 
@@ -9,15 +8,13 @@ import SpotifyMusicPlayer from "./SpotifyPlayer"
 
 function MusicContent() {
     // var musicid = useSelector((state) => state.spotify.clientid)
-    const showContent = useSelector((state) => state.controls.showContentBox)
-    const musicToken = useSelector((state) => state.spotify.accesstoken)
-    const musicPlaylist = useSelector((state) => state.content.musicPlaylistCodes)
+    // const showContent = useSelector((state) => state.controls.showContentBox)
+    // const musicToken = useSelector((state) => state.spotify.accesstoken)
     // const userMusic = useSelector((state) => state.content.musicButtonChoice)
-    const dispatch = useDispatch()
+
 
     // console.log(`spotify current search keyword from button is: ${userMusic}`)
-    console.log(`spotify current token is: ${musicToken}`)
-
+    // console.log(`spotify current token is: ${musicToken}`)
 
     // const spotifyApi = new SpotifyWebApi({
     //     clientId: `${musicid}`,
@@ -29,7 +26,6 @@ function MusicContent() {
     //     if (musicToken === "none token") return
     //     spotifyApi.setAccessToken(musicToken)
     // }, [musicToken])
-
 
     // const musicDummyID = `1DFixLWuPkv3KT3TnV35m3`
 
@@ -50,25 +46,16 @@ function MusicContent() {
     //     console.log('Something went wrong!', err);
     // })
 
-
-    function contentToggleHandler() {
-        dispatch(controlActions.toggleContentBox())
-    }
-
-    var currentPlaylist = 0
-
     // console.log(userMusicContent)
 
-    return (
-        <div className="musiccontent">
-            {/* <div>
-                <button onClick={contentToggleHandler}>Toggle music on/off</button>
-            </div> */}
 
-            {showContent && <div>
+
+    return (
+        <div className="card bg-success rounded ">
+            <div className="card-body">
                 <SpotifyMusicPlayer />
                 {/* <iframe title="playlist" src={`https://open.spotify.com/embed/playlist/${musicPlaylist[(currentPlaylist)]}`} width="300" height="380" frameborder="0" allowtransparency="true" allow="encrypted-media"></iframe> */}
-            </div>}
+            </div>
         </div>
     )
 }
