@@ -1,19 +1,15 @@
 import MusicButton from "./MusicButton"
 import { useEffect } from "react"
-import { useDispatch } from "react-redux"
 import { useSelector } from "react-redux"
 import SpotifyWebApi from "spotify-web-api-node"
-import { contentActions } from "../store/Content"
 
-var userMusicContent = `pending`
+
 
 function MusicForm(props) {
     var musicid = useSelector((state) => state.spotify.clientid)
-    const musicToken = useSelector((state) => state.spotify.accesstoken)
+    const musicToken = localStorage.getItem("spotifyToken")
     const musicChoice = useSelector((state) => state.content.musicButtonChoice)
-    const musicPlaylist = useSelector((state) => state.content.musicPlaylistCodes)
 
-    const dispatch = useDispatch()
 
     const spotifyApi = new SpotifyWebApi({
         clientId: `${musicid}`,
@@ -28,49 +24,65 @@ function MusicForm(props) {
 
 
     function updateContent(content) {
-        localStorage.setItem("playlistCodes", content[0])
+        for (let i = 0; i < 10; i++) {
+            localStorage.setItem(`musicplaylistcode${i}`, `${content[i]}`)
+        }
+    }
+
+    if (musicChoice === "none") {
+        spotifyApi.searchPlaylists("peaceful").then(
+            (res) => {
+                console.log(`music playlist res:`, res.body)
+                var playlistMusicItems = res.body.playlists.items
+                var playlistCodes = playlistMusicItems.map(playlist => {
+                    return playlist.uri
+                })
+                updateContent(playlistCodes)
+            }
+        ).catch((err) => {
+            console.log('Something went wrong!', err);
+        })
+    } else {
+        spotifyApi.searchPlaylists(`${musicChoice}`).then(
+            (res) => {
+                console.log(`music playlist res:`, res.body)
+                var playlistMusicItems = res.body.playlists.items
+                var playlistCodes = playlistMusicItems.map(playlist => {
+                    return playlist.uri
+                })
+                updateContent(playlistCodes)
+            }
+        ).catch((err) => {
+            console.log('Something went wrong!', err);
+        })
     }
 
 
 
-    spotifyApi.searchPlaylists(`${musicChoice}`).then(
-        (res) => {
-            //save to an array?
-            console.log(`playlist res:`, res.body)
-            var playlistMusic = res.body.playlists.items
-            var playlistCodes = playlistMusic.map(playlist => {
-                return playlist.uri
-            })
-            updateContent(playlistCodes)
-        }
-    ).catch((err) => {
-        console.log('Something went wrong!', err);
-    })
 
-    useEffect(() => {
-        console.log("new playlist changes", userMusicContent)
-        console.log("current playlist state", musicPlaylist)
 
-        dispatch(contentActions.updatePlaylists(userMusicContent))
-    }, [userMusicContent])
 
 
 
 
     return (
-        <div className="pracmusicmenu bg-dark">
-            <div>
-                <h2>Music Choices</h2>
-            </div>
-            <div>
-                <p>{`Your current music choice: ${musicChoice}`}</p>
-            </div>
-            <div className="menubuttons">
-                <MusicButton musicKeyword="Piano" />
-                <MusicButton musicKeyword="Jazz" />
-                <MusicButton musicKeyword="Acoustic" />
-                <MusicButton musicKeyword="Lofi" />
-                <MusicButton musicKeyword="Yoga" />
+        <div className="card bg-dark m-4">
+            <div className="card-body">
+                <div className="d-flex flex-column justify-content-center">
+                    <div className="card-title">
+                        <h2>Music Choices</h2>
+                    </div>
+                    <div className="card-text">
+                        <p>{`Your current music choice: ${musicChoice}`}</p>
+                    </div>
+                    <div className="d-flex flex-column justify-content-center">
+                        <MusicButton musicKeyword="Piano" />
+                        <MusicButton musicKeyword="Coffee Shop" />
+                        <MusicButton musicKeyword="Acoustic" />
+                        <MusicButton musicKeyword="Lofi" />
+                        <MusicButton musicKeyword="Yoga" />
+                    </div>
+                </div>
             </div>
         </div>
     )
