@@ -11,16 +11,14 @@ const buttonStyles = {
 function AtmosNavbar() {
 
     return (
-        <nav className="bg-dark">
-            <div className="container-fluid">
-            <img alt="menu" src={title}></img>
-                <div className="navbarbtn d-flex justify-content-end">
-                    <li>
-                        <Link to="/menu" ><img alt="menu" style={buttonStyles} src={menu}></img></Link>
-                    </li>
-                    <li>
-                        <Link to="Dashboard"><img alt="room" src={room}></img></Link>
-                    </li>
+        <nav className="bg-primary">
+            <div className="container-fluid d-flex justify-content-between">
+                <div className="pb-4">
+                    <img alt="menu" src={title}></img>
+                </div>
+                <div className="navbarbtn d-flex align-self-center">
+                    <Link to="/menu" ><img alt="menu" style={buttonStyles} src={menu}></img></Link>
+                    {/* <Link to="/dashboard"><img alt="room" src={room}></img></Link> */}
                 </div>
             </div>
         </nav>)
