@@ -14,30 +14,25 @@ function Welcome(props) {
 
     return (
         <div className={styles.bg}>
-            <div className={styles.align}>
-                <div>
-                    <p>anything</p>
+            <div className={styles.containerdiv}>
+                <img alt="" className="img-fluid" src={title}></img>
+                <div className={styles.drop}>
+                    <img className="img-fluid mb-3" alt="" src={tv}></img>
                 </div>
-                <div className={styles.working}>
-                    <img alt="" className="img-fluid" src={title}></img>
-                    <div className={styles.drop}>
-                        <img className="img-fluid mb-3" alt="" src={tv}></img>
-                    </div>
+                <div className={styles.containerdivintro}>
                     <p className={styles.text} >A Visual Auditory Experience</p>
                     <AnimationWrapper>
-                        <div >
-                            {!isAuth && (
-                                <div>
-                                    <p className={styles.texts} >*NOTE* For this app to work we need to use some of your data from a Premium Spotify account! Please note we do note keep any of the data you lend, it is only to grant access to spotify features.</p>
-                                    <button className="btn btn-success" onClick={props.userWelcome}>Agree</button>
-                                </div>)}
-                            {isAuth && (<div>
-                                <p className={styles.texts} >*NOTE* Thank you! You will need to be redirected to the spotify website in order to obtain your credentials, you will be redirected back here once you're done!</p>
-                                <button className="btn btn-success" onClick={props.onRedirect}>Authorize on Spotify</button>
+                        {!isAuth && (
+                            <div className={styles.containerdiv}>
+                                <p className={styles.texts} >*NOTE* For this app to work we need to use some of your data from a Premium Spotify account! Please note we do note keep any of the data you lend, it is only to grant access to spotify features.</p>
+                                <button className="m-3 btn btn-success" onClick={props.userWelcome}>Agree</button>
                             </div>)}
-                        </div>
+                        {isAuth && (<div className={styles.containerdiv}>
+                            <p className={styles.texts} >*NOTE* Thank you! You will need to be redirected to the spotify website in order to obtain your credentials, you will be redirected back once you're done!</p>
+                            <button className="m-3 btn btn-success" onClick={props.onRedirect}>Authorize on Spotify</button>
+                        </div>)}
                     </AnimationWrapper>
-                </div >
+                </div>
             </div >
         </div>
     )
