@@ -4,71 +4,47 @@ import { AnimationWrapper } from 'react-hover-animation'
 import title from '../images/title.png';
 import welcome from '../images/welcome.png';
 import styles from './welcomeStyles.module.css'
+import { Link } from 'react-router-dom';
 
-// const buttonStyles = {
-//     display: 'flex',
-//     textAlign: "center",
-//     margin: "auto",
-//     cursor: 'pointer',
-//     padding:'0.46em 1.6em',
-// 	border:'0.1em solid #FFFFFF',
-// 	borderRadius:'0.12em',
-// 	boxSizing: 'border-box',
-// 	textDecoration:'none',
-// 	fontFamily:'OCR A Std',
-// 	fontWeight:300,
-// 	color:'#FFFFFF',
-// 	textShadow: '0 0.04em 0.04em rgba(0,0,0,0.35)',
-// 	backgroundColor:'#7fb069',
-   
-// }
 
-// const drop = {
-//     filter: 'drop-shadow(40px 20px 12px )',
-//     display: 'flex',
-// justifyContent: 'center'
-// }
-
-// const align = {
-//     display: 'flex',
-//     justifyContent: 'center'
-//     }
-// const styles = {
-//     margin: 'none',
-//     background: '#7fb069',
-//     height: '100vh',
-//     width: '100%'
-// }
 
 function WelcomeBack(props) {
+
+    var authCode = props.newCode
+    console.log(authCode)
+
+
+    localStorage.setItem("spotifyCode", `${authCode}`)
+
+    useAuth(authCode)
+
+
     const isAuth = useSelector((state) => state.spotify.isSpotifyAuth)
     var accesstoken = useSelector((state) => state.spotify.accesstoken)
     console.log(`welcome back comp loaded is auth ${isAuth}`)
     console.log(`welcome back comp loaded is current access token ${accesstoken}`)
-
     localStorage.setItem("spotifyToken", `${accesstoken}`)
-
-    var code = props.newCode
-    useAuth(code)
-
-
 
 
     return (
         <div className={styles.bg}>
-        <div  className={styles.align}>
-            <div >
-           
-            <div className={styles.drop}>
-    <img className="img-fluid" alt="" src={welcome}></img>
-    </div>
-            <AnimationWrapper>
-                        <button className={styles.buttonStyles} onClick={props.onContinue}>Enter</button>
-            </AnimationWrapper>
-        </div>
-        </div>
-        </div>
-       
+            <div className={styles.containerdiv}>
+                <div className={styles.drop}>
+                    <img className="img-fluid" alt="" src={welcome}></img>
+                </div>
+                <div className="card text-center">
+                    <div className="card-body">
+                        <h5 className="card-title">Welcome to atmos! </h5>
+                        <p className="card-text">While using the app we combine youtube and spotify to allow you to both listen and watch anything of your choosing! Click whichever genre you would like to enjoy and relax!</p>
+                        <div>
+                            <AnimationWrapper>
+                                <Link to="/menu" className="m-3 btn btn-success">Continue</Link>
+                            </AnimationWrapper>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div >
     )
 }
 
