@@ -3,8 +3,7 @@ import { createSlice } from "@reduxjs/toolkit"
 
 // setting the initail state object, this is the master state which will be an expanded object
 const initialContentState = {
-    musicButtonChoice: "no music",
-    musicPlaylistCodes: ["no playlist"],
+    musicButtonChoice: "none",
     showCreateModal: false,
 }
 
@@ -15,9 +14,6 @@ const contentSlice = createSlice({
     reducers: {
         updateMusicContent(state, action) {
             state.musicButtonChoice = action.payload
-        },
-        updatePlaylists(state, action) {
-            state.musicPlaylistCodes = action.payload
         },
 
         showCreateModal(state) {
