@@ -23,7 +23,6 @@ const initialYoutubeState = {
             "https://www.youtube.com/watch?v=21HKaqA1rpo",
             "https://www.youtube.com/watch?v=JB0A8Me8EKk",
             "https://www.youtube.com/watch?v=CcDJZ5_nGZs",
-            ""
         ],
         Forest: [
             "https://www.youtube.com/watch?v=qRTVg8HHzUo",
@@ -31,17 +30,16 @@ const initialYoutubeState = {
             "https://www.youtube.com/watch?v=hld4uaO1MDE",
             "https://www.youtube.com/watch?v=478TeAxm12g",
             "https://www.youtube.com/watch?v=qQU0UD8m6pE",
-            "https://www.youtube.com/watch?v=-tMVBF78RMU"
         ],
         Travel: [
             "https://www.youtube.com/watch?v=P6CLmbAwvMk",
             "https://www.youtube.com/watch?v=fh3EdeGNKus",
             "https://www.youtube.com/watch?v=vXCB1zGGFiY",
-            "https://www.youtube.com/watch?v=ADt_RisXY0U"
+            "https://www.youtube.com/watch?v=ADt_RisXY0U",
+            "https://www.youtube.com/watch?v=V_vj7BMawAI"
         ],
     },
-    currentChoice: "no video",
-    currentContent: [],
+    currentChoice: "none",
 
 }
 
